+++ neo4j_runway/resources/prompts/data_modeling/initial_data_model.py
@@ -1,43 +1,12 @@
 from typing import Any, Dict, Optional
 
-from ...llm_response_types.initial_model_pool import DataModelEntityPool
 from .constants import (
     DATA_MODEL_FORMAT,
-    ENTITY_POOL_GENERATION_RULES,
     NODE_GENERATION_RULES,
     NODES_FORMAT,
 )
 from .formatters import get_rules
 from .template import create_data_modeling_prompt
-
-
-def create_initial_data_model_cot_prompt(
-    discovery_text: str,
-    multifile: bool,
-    use_cases: Optional[str],
-    valid_columns: Dict[str, Any],
-    data_dictionary: Optional[Dict[str, Any]] = None,
-) -> str:
-    """
-    Generate a prompt to find nodes, relationships and properties to include in a data model.
-    This is only for brainstorming, result of prompt should not be a DataModel.
-
-    Returns
-    -------
-    str
-        The prompt.
-    """
-    prefix = "Please generate a pool of entities that will be used to construct a graph data model."
-
-    return create_data_modeling_prompt(
-        prefix=prefix,
-        discovery=discovery_text,
-        multifile=multifile,
-        use_cases=use_cases,
-        valid_columns=valid_columns,
-        data_dictionary=data_dictionary,
-        rules=ENTITY_POOL_GENERATION_RULES,
-    )
 
 
 def create_initial_nodes_prompt(
@@ -88,7 +57,6 @@ def create_initial_data_model_prompt(
     return create_data_modeling_prompt(
         prefix=prefix,
         discovery=discovery_text,
-        entity_pool=data_model_recommendations,
         valid_columns=valid_columns,
         data_dictionary=data_dictionary,
         use_cases=use_cases,