+++ neo4j_runway/resources/prompts/data_modeling/template.py
@@ -31,7 +31,6 @@ def create_data_modeling_prompt(
     errors: Optional[List[str]] = None,
     corrections: Optional[str] = None,
     data_model: Optional["DataModel"] = None,  # type: ignore
-    entity_pool: Optional["DataModelEntityPool"] = None,  # type: ignore
     nodes: Optional["Nodes"] = None,  # type: ignore
     use_cases: Optional[str] = None,
     data_model_format: Optional[str] = None,
@@ -52,8 +51,6 @@ def create_data_modeling_prompt(
         res += format_corrections(corrections=corrections)
     if data_model is not None:
         res += format_data_model(data_model=data_model, yaml_format=data_model_as_yaml)
-    if entity_pool is not None:
-        res += format_entity_pool(entity_pool=entity_pool, retry_prompt=retry_prompt)
     if nodes is not None:
         res += format_nodes(nodes=nodes, retry_prompt=retry_prompt)
     if use_cases is not None: